Extract day calendar view
