+++ lib/ui/schedule/view_model/day_viewmodel.dart
@@ -1,0 +1,5 @@
+import 'package:stacked/stacked.dart';
+
+class DayViewModel extends BaseViewModel {
+
+}
