+++ lib/ui/schedule/widgets/schedule_view.dart
@@ -6,10 +6,11 @@ import 'package:flutter/services.dart';
 import 'package:calendar_view/calendar_view.dart' as calendar_view;
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:intl/intl.dart';
+import 'package:notredame/ui/schedule/widgets/calendars/day_calendar.dart';
 import 'package:notredame/ui/schedule/widgets/calendars/month_calendar.dart';
 import 'package:notredame/ui/schedule/widgets/calendars/week_calendar.dart';
+import 'package:notredame/ui/schedule/widgets/day_view_header.dart';
 import 'package:stacked/stacked.dart';
-import 'package:table_calendar/table_calendar.dart';
 
 // Project imports:
 import 'package:notredame/data/services/analytics_service.dart';
@@ -17,13 +18,10 @@ import 'package:notredame/data/services/calendar_service.dart';
 import 'package:notredame/data/services/signets-api/models/course_activity.dart';
 import 'package:notredame/domain/constants/preferences_flags.dart';
 import 'package:notredame/locator.dart';
-import 'package:notredame/ui/core/themes/app_palette.dart';
-import 'package:notredame/ui/core/themes/app_theme.dart';
 import 'package:notredame/ui/core/ui/base_scaffold.dart';
 import 'package:notredame/ui/dashboard/widgets/course_activity_tile.dart';
 import 'package:notredame/ui/schedule/view_model/schedule_viewmodel.dart';
 import 'package:notredame/ui/schedule/widgets/calendar_selector.dart';
-import 'package:notredame/ui/schedule/widgets/schedule_calendar_tile.dart';
 import 'package:notredame/ui/schedule/widgets/schedule_settings.dart';
 
 class ScheduleView extends StatefulWidget {
@@ -47,26 +45,6 @@ class _ScheduleViewState extends State<ScheduleView>
   final AnalyticsService _analyticsService = locator<AnalyticsService>();
 
   static const String tag = "ScheduleView";
-  static bool _isDayViewAnimating = false;
-
-  late AnimationController _animationController;
-
-  @override
-  void initState() {
-    super.initState();
-    _animationController = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 400),
-    );
-
-    _animationController.forward();
-  }
-
-  @override
-  void dispose() {
-    _animationController.dispose();
-    super.dispose();
-  }
 
   @override
   Widget build(BuildContext context) =>
@@ -103,7 +81,7 @@ class _ScheduleViewState extends State<ScheduleView>
       return WeekCalendar(model: model, eventController: eventController, weekViewKey: weekViewKey);
     }
     if (!model.calendarViewSetting) {
-      return _buildCalendarViewDaily(model, context, eventController);
+      return DayCalendar(m: model, eventController: eventController, dayViewKey: dayViewKey);
     }
 
     return _buildListView(model, context, eventController);
@@ -128,7 +106,7 @@ class _ScheduleViewState extends State<ScheduleView>
         child: ListView(
           padding: EdgeInsets.zero,
           children: [
-            _buildTableCalendar(model, eventController),
+            DayViewHeader(m: model, eventController: eventController, dayViewKey: dayViewKey),
             const SizedBox(height: 8.0),
             const Divider(indent: 8.0, endIndent: 8.0, thickness: 1),
             const SizedBox(height: 6.0),
@@ -150,208 +128,11 @@ class _ScheduleViewState extends State<ScheduleView>
     ]);
   }
 
-  Widget _buildCalendarViewDaily(ScheduleViewModel model, BuildContext context,
-      calendar_view.EventController eventController) {
-    final double heightPerMinute =
-        (MediaQuery.of(context).size.height / 1200).clamp(0.45, 1.0);
-    return Column(children: [
-      _buildTableCalendar(model, eventController),
-      Expanded(
-          child: calendar_view.DayView(
-              showVerticalLine: false,
-              dayTitleBuilder: calendar_view.DayHeader.hidden,
-              key: dayViewKey,
-              controller: eventController
-                ..addAll(model.selectedDayCalendarEvents()),
-              onPageChange: (date, page) => setState(() {
-                    if (!_isDayViewAnimating) {
-                      model.handleViewChanged(date, eventController, []);
-                    }
-                  }),
-              backgroundColor: context.theme.scaffoldBackgroundColor,
-              initialDay: model.daySelected,
-              // height occupied by 1 minute time span.
-              hourIndicatorSettings: calendar_view.HourIndicatorSettings(
-                color: context.theme.appColors.scheduleLine,
-              ),
-              liveTimeIndicatorSettings:
-                  calendar_view.LiveTimeIndicatorSettings(
-                color: context.theme.textTheme.bodyMedium!.color!,
-              ),
-              heightPerMinute: heightPerMinute,
-              scrollOffset: heightPerMinute * 60 * 7.5,
-              keepScrollOffset: true,
-              dateStringBuilder: (date, {secondaryDate}) {
-                final locale = AppIntl.of(context)!.localeName;
-                return '${date.day} ${DateFormat.MMMM(locale).format(date)} ${date.year}';
-              },
-              timeStringBuilder: (date, {secondaryDate}) {
-                return DateFormat('H:mm').format(date);
-              },
-              eventTileBuilder:
-                  (date, events, boundary, startDuration, endDuration) =>
-                      _buildEventTile(events, context))),
-    ]);
-  }
-
-  Widget _buildEventTile(
-    List<calendar_view.CalendarEventData<dynamic>> events,
-    BuildContext context,
-  ) {
-    if (events.isNotEmpty) {
-      return ScheduleCalendarTile(
-        title: events[0].title,
-        description: events[0].description,
-        start: events[0].startTime,
-        end: events[0].endTime,
-        titleStyle: TextStyle(
-          fontSize: 14,
-          color: events[0].color.accent,
-        ),
-        totalEvents: events.length,
-        padding: const EdgeInsets.all(7.0),
-        backgroundColor: events[0].color,
-        borderRadius: BorderRadius.circular(6.0),
-        buildContext: context,
-      );
-    } else {
-      return Container();
-    }
-  }
-
   Widget _buildTitleForDate(DateTime date, ScheduleViewModel model) => Center(
           child: Text(
         DateFormat.MMMMEEEEd(model.locale.toString()).format(date),
         style: Theme.of(context).textTheme.headlineMedium,
       ));
-
-  /// Build the square with the number of [events] for the [date]
-  Widget? _buildEventsMarker(
-      ScheduleViewModel model, DateTime date, List events, Color color) {
-    if (events.isNotEmpty) {
-      return Positioned(
-        right: 1,
-        bottom: 1,
-        child: AnimatedContainer(
-          duration: const Duration(milliseconds: 200),
-          decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(5), color: color),
-          width: 16.0,
-          height: 16.0,
-          child: Center(
-            child: Text(
-              '${events.length}',
-              style: const TextStyle().copyWith(
-                fontSize: 12.0,
-              ),
-            ),
-          ),
-        ),
-      );
-    }
-    return null;
-  }
-
-  /// Build the calendar
-  Widget _buildTableCalendar(
-      ScheduleViewModel model, calendar_view.EventController eventController) {
-    const Color selectedColor = AppPalette.etsLightRed;
-    final Color todayColor = context.theme.appColors.dayIndicatorWeekView;
-    final Color defaultColor = context.theme.appColors.scheduleLine;
-
-    return TableCalendar(
-      key: const Key("TableCalendar"),
-      locale: model.locale?.toLanguageTag(),
-      selectedDayPredicate: (day) {
-        return isSameDay(model.daySelected, day);
-      },
-      headerStyle: HeaderStyle(
-          titleTextFormatter: (date, locale) =>
-              DateFormat.MMMMEEEEd(locale).format(model.daySelected),
-          titleCentered: true,
-          formatButtonVisible: false),
-      eventLoader: model.coursesActivitiesFor,
-      calendarFormat: CalendarFormat.week,
-      focusedDay: model.daySelected,
-      onPageChanged: (focusedDay) {
-        // Used to compare
-        model.listViewCalendarSelectedDate = focusedDay;
-      },
-      calendarBuilders: CalendarBuilders(
-          defaultBuilder: (context, date, _) =>
-              _buildSelectedDate(date, defaultColor, model, eventController),
-          outsideBuilder: (context, date, _) =>
-              _buildSelectedDate(date, defaultColor, model, eventController),
-          todayBuilder: (context, date, _) =>
-              _buildSelectedDate(date, todayColor, model, eventController),
-          selectedBuilder: (context, date, _) => FadeTransition(
-                opacity:
-                    Tween(begin: 0.0, end: 1.0).animate(_animationController),
-                child: _buildSelectedDate(
-                    date, selectedColor, model, eventController),
-              ),
-          markerBuilder: (context, date, events) {
-            final bool isSelected = isSameDay(date, model.daySelected);
-            final bool isToday = isSameDay(date, DateTime.now());
-            final Color color = isSelected
-                ? selectedColor
-                : isToday
-                    ? todayColor
-                    : defaultColor;
-            return _buildEventsMarker(model, date, events, color);
-          }),
-      // Those are now required by the package table_calendar ^3.0.0. In the doc,
-      // it is suggest to set them to values that won't affect user experience.
-      // Outside the range, the date are set to disable so no event can be loaded.
-      firstDay: DateTime.utc(2010, 12, 31),
-      lastDay: DateTime.utc(2100, 12, 31),
-    );
-  }
-
-  /// Build the visual for the selected [date]. The [color] parameter set the color for the tile.
-  Widget _buildSelectedDate(DateTime date, Color color, ScheduleViewModel model,
-          calendar_view.EventController eventController) =>
-      Container(
-        margin: const EdgeInsets.all(4.0),
-        decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(10),
-            border: Border.all(color: color)),
-        child: InkWell(
-          borderRadius: BorderRadius.circular(10),
-          onTap: () {
-            if (dayViewKey.currentState != null) {
-              _isDayViewAnimating = true;
-            }
-
-            setState(() => model.handleViewChanged(date, eventController, []));
-
-            dayViewKey.currentState
-                ?.animateToDate(model.daySelected)
-                .then((value) => _isDayViewAnimating = false);
-          },
-          child: Container(
-            padding: const EdgeInsets.only(top: 5.0, left: 6.0),
-            width: 100,
-            height: 100,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  '${date.day}',
-                  style: const TextStyle().copyWith(
-                    fontSize: 16.0,
-                    height: 1.2,
-                  ),
-                ),
-                if (date.month != DateTime.now().month ||
-                    date.year != DateTime.now().year)
-                  Text(DateFormat.MMM(model.locale.toString()).format(date),
-                      style: const TextStyle(fontSize: 10.0)),
-              ],
-            ),
-          ),
-        ),
-      );
 
   /// Build the list of the events for the selected day.
   Widget _buildEventList(List<dynamic> events) {
@@ -378,8 +159,7 @@ class _ScheduleViewState extends State<ScheduleView>
             );
           },
         ),
-        if ((model.settings[PreferencesFlag.scheduleShowTodayBtn] as bool) ==
-            true)
+        if ((model.settings[PreferencesFlag.scheduleShowTodayBtn] as bool))
           IconButton(
               icon: const Icon(Icons.today_outlined),
               onPressed: () => setState(() {
